separate ui & server helper fns
